Desarrollo

+++ commands/Pure/uwu.js
@@ -1,5 +1,5 @@
-const { rand } = require('../../func');
-const { CommandOptionsManager, CommandMetaFlagsManager, CommandManager } = require('../Commons/commands');
+const { randInArray } = require('../../func');
+const { CommandOptions, CommandTags, CommandManager } = require('../Commons/commands');
 
 const uwusopt = [
 	'<:uwu:681935702308552730>',
@@ -9,11 +9,11 @@ const uwusopt = [
 	'<:uwu5:720506981743460472>'
 ];
 
-const options = new CommandOptionsManager()
+const options = new CommandOptions()
 	.addParam('uwu', { name: 'uwu', expression: 'uwu' }, 'uwu', { optional: true })
 	.addFlag('u', 'uwuwu', 'uwu')
 	.addFlag('bd', ['borrar', 'delete'], 'para borrar el mensaje original');
-const flags = new CommandMetaFlagsManager().add(
+const flags = new CommandTags().add(
 	'MEME',
 	'EMOTE',
 );
@@ -21,14 +21,19 @@ const command = new CommandManager('uwu', flags)
 	.setAliases('uwu')
 	.setLongDescription('uwu')
 	.setOptions(options)
-	.setExecution(async (request, args, isSlash) => {
-		const deleteOriginal = !isSlash && options.fetchFlag(args, 'borrar');
-		const randomUwu = uwusopt[rand(uwusopt.length)];
+	.setExperimental(true)
+	.setExperimentalExecution(async (request, args) => {
+		const deleteOriginal = args.parseFlag('borrar');
+		const randomUwu = randInArray(uwusopt);
 		
-		if(deleteOriginal)
-			request.delete().catch(console.error)
-		
-		return request.reply({ content: randomUwu });
+		if(deleteOriginal && args.isMessageSolver()) {
+			return Promise.all([
+				request.channel.send({ content: randomUwu }).catch(console.error),
+				request.delete().catch(console.error),
+			]);
+		}
+
+		return request.reply({ content: randomUwu }).catch(console.error);
 	});
 
 module.exports = command;
